+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### :bug: Fixed
 
 - Crash when handling connection errors with empty header values.
+- Fuzzing phase not generating format-violating values for custom string formats.
 
 ## [4.6.3](https://github.com/schemathesis/schemathesis/compare/v4.6.2...v4.6.3) - 2025-11-26